requirements.txt
cycler==0.10.0
decorator==4.0.11
ipython==6.1.0
ipython-genutils==0.2.0
jedi==0.10.2
line-profiler==2.0
matplotlib==2.0.2
memory-profiler==0.47
numpy==1.13.0
olefile==0.44
pandas==0.20.2
pexpect==4.2.1
pickleshare==0.7.4
Pillow==4.2.0
pkg-resources==0.0.0
prompt-toolkit==1.0.14
ptyprocess==0.5.2
Pygments==2.2.0
pyparsing==2.2.0
python-dateutil==2.6.0
pytz==2017.2
scikit-learn==0.18.2
scipy==0.19.1
simplegeneric==0.8.1
six==1.10.0
tqdm==4.14.0
traitlets==4.3.2
wcwidth==0.1.7
bhtsne
Cython
